Crate mega

Structs

Client

A client

EasyClient

A client

EasyFileDownloadReader

A reader for a file.

EasyGetAttributesBuilder

A builder for a get_attributes call.

ErrorCode

An API Error

FetchNodesResponse

FetchNodes command response

FileAttributes

File attributes

FileKey

The encryption key for a file.

FileValidationError

An error that occurs when a file fails validation.

FileValidator

A struct to validate files.

FolderKey

The encryption key for a folder.

GetAttributesResponse

GetAttributes command response

ParsedMegaFileUrl

A parsed file url

ParsedMegaFolderUrl

A parsed folder url

Url

A parsed URL record.

Enums

Command

A command

DecodeAttributesError

An error that may occur while decoding attributes

Error

The library error type

FetchNodesNodeKind

The kind of node

FileKeyParseError

An error that may occur while parsing a FileKey.

FileOrFolderKey

Either a file or folder key

FolderKeyParseError

An error that may occur while parsing a FolderKey.

ParseMegaUrlError

An error that may occur while parsing a mega url.

ParsedMegaUrl

A parsed mega url.

Response

An api response

ResponseData

API Response data